gen3-integration-tests/pyproject.toml
[tool.poetry]
name = "gen3-integration-tests"
version = "0.1.0"
description = "Gen3 Integration Tests"
authors = ["CTDS UChicago <cdis@uchicago.edu>"]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/uc-cdis/gen3-code-vigil"
include = [
  "NOTICE",
]

[tool.poetry.dependencies]
python = "^3.9"
allure-pytest = "^2.13.2"
boto3 = "~1.28.33"
cdislogging = "<2"
data-simulator = {git = "https://github.com/uc-cdis/data-simulator", rev = "1.6.1"}
filelock = "<4"
gen3 = "<5"
pytest = "^7.4.0"
pytest-datadir = "^1.4.1"
pytest-env = "^1.0.0"
pytest-html = "^3.2.0"
pytest-md-report = "^0.5.0"
pytest-playwright = "^0.4.3"
pytest-xdist = "^3.3.1"
python-dotenv = "^1.0.0"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--log-cli-level=ERROR"
markers = [
  # general
  "wip: marks tests as `Work In Progress` and skips them in CI pipeline",
  # services
  "audit: tests for audit-service",
  "fence: test for fence service",
  "requestor: tests for requestor service",
  "indexd: tests for indexd service",
  "manifestservice: tests for manifest service",
  "mds: tests for metadata-service",
  "portal: tests for data-portal",
  "frontend_framework: tests for frontend-framework",
  "gen3_client: tests for gen3-client / cdis-data-client",
  "sheepdog: tests for sheepdog",
  "peregrine: tests for peregrine",
  # features
  "agg_mds: tests for aggregate metadata service",
  "workspace: tests for workspace page",
  "client_credentials: tests for fence client credentials",
  "graph_submission: tests for submitting graph data",
  "graph_query: tests for querying graph data",
  "guppy: tests for guppy-service",
  "audit: tests for audit-service",
  "fence: test for fence service",
  "requestor: tests for requestor service",
  "indexd: tests for indexd service",
  "manifestservice: tests for manifest service"
]
pythonpath = "."
md_report = "true"
md_report_verbose = 0
md_report_flavor = "gfm"
md_report_output = "output/report.md"
